requirements.txt
anyio==3.7.0
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
dill==0.3.6
distro==1.8.0
dnspython==2.3.0
docker==6.1.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
exceptiongroup==1.1.1
fastapi==0.98.0
Flask==2.3.2
Flask-Cors==4.0.0
fonttools==4.40.0
h11==0.14.0
httptools==0.5.0
idna==3.4
imbalanced-learn==0.10.1
imblearn==0.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.0
jsonschema==3.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
mypy-boto3-s3==1.26.163
numpy==1.24.4
packaging==23.1
pandas==2.0.3
paramiko==3.2.0
Pillow==9.5.0
pip-chill==1.0.3
pycparser==2.21
pydantic==1.10.9
pymongo==4.4.0
PyNaCl==1.5.0
pyparsing==3.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3
pywin32==306
PyYAML==5.4.1
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
texttable==1.6.7
threadpoolctl==3.1.0
types-awscrt==0.16.21
types-s3transfer==0.6.1
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
watchfiles==0.19.0
websocket-client==0.59.0
websockets==11.0.3
Werkzeug==2.3.6
wincertstore==0.2
zipp==3.15.0
-e .